requirements.txt
flask==3.1.1
numpy==1.23.5
azure-storage-blob==12.19.1
pandas==1.5.3
cloudpickle==3.1.1
scikit-learn==1.3.0
gunicorn